code/extract/requirements.txt
google-cloud-storage==1.42.2